requirements.txt
pure-python-adb==0.3.0.dev0
pillow==10.0.1
pytesseract==0.3.10
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
selenium==4.15.2
beautifulsoup4==4.12.2
